fix: 1차 코드 분리
폴더구조에 맞게 코드 분리
