+++ src/basic/components/CartList.js
@@ -1,0 +1,6 @@
+export const CartList = () => {
+  let cartList = document.createElement("select");
+  cartList.id = "product-select";
+  cartList.className = "border rounded p-2 mr-2";
+  return cartList;
+};
