+++ src/basic/components/HeaderTitle.js
@@ -1,0 +1,6 @@
+export const HeaderTitle = () => {
+  let headerTitle = document.createElement("h1");
+  headerTitle.className = "text-2xl font-bold mb-4";
+  headerTitle.textContent = "장바구니";
+  return headerTitle;
+}
