+++ src/basic/components/StockInformation.js
@@ -1,0 +1,6 @@
+export const StockInformation = () => {
+  let stockInfo = document.createElement("div");
+  stockInfo.id = "stock-status";
+  stockInfo.className = "text-sm text-gray-500 mt-2";
+  return stockInfo;
+};
