+++ src/basic/components/TotalPrice.js
@@ -1,0 +1,6 @@
+export const TotalPrice = () => {
+  let totalPrice = document.createElement("div");
+  totalPrice.id = "cart-total";
+  totalPrice.className = "text-xl font-bold my-4";
+  return totalPrice;
+};
